requirements.txt
Django==1.6.1
django-model-utils==2.0
djangorestframework==2.3.12
wsgiref==0.1.2
